Fix segfault with relative argv[0] in client
Running `LD_PRELOAD=/path/to/librm2fb_client.so.1.0.0 xochitl` currently
results in a segfault because `swtfb::locate_signature` is called with a
relative binary path, and therefore `fopen` fails in `swtfb::read_file`.

This patch replaces the use of `argv[0]` with reading `/proc/self/exe`
which always return an absolute path to the current executable.

It also adds some error checking to `swtfb::read_file` so that we get a
nicer error message instead of a segfault, should a similar bug happen
again.