Do not export unused symbols in librm2fb_client.so (#41)
All symbols from the libfrida-gum.a static library were previously
re-exported in librm2fb_client.so, plus some symbols in the client
source that are not needed externally. This patch reduces the exported
symbols to a bare minimum, which limits the risk of conflicts with
binaries that the library is LD_PRELOAD-ed into.