Add support for release 2.9
This PR replaces our existing strategies for dynamically locating the various functions we need to make rm2fb work with a static lookup table of the addresses of these functions.

We’re making this choice because it seems that we’re spending more time trying to adapt the strategies after each release than we would if we just updated a static list of addresses. This change also gives us the opportunity to check that rm2fb works correctly with an OS release before users get to use it. The cost is that we’ll need to make an update after each OS release.

To test, you’ll need to move the `offsets` file in the `src/` folder to `/opt/share/rm2fb/offsets` on the tablet (the parent directory needs to be created). At the moment, that file only contains the offsets for the latest release (2.9.0.153).