Update timespec overflow implementation
This was raised by @fenollp when reviewing the rust-impl of this client.

https://github.com/canselcik/libremarkable/pull/78#discussion_r748864139

The code wouldn't work properly if SEM_WAIT_TIMEOUT would be greater
than one second, since only one overflow is accounted for with an "if".

Using modulo here instead of while so no branching is needed and it
would scale linearly not not require one loop per overflown second.